Update build.rs
u128 doesn't behave well over FFI and generates warnings for its use.
The only things using 128-bit ints are libc math functions, which don't
need to be exported here at all. This change creates a block list and we
use it to make sure these math functions are not included.

We want to keep the "not FFI safe" warning enabled in case it catches a
real issue.